refactor: use CML utilities were applicable
