+++ src/plugin.js
@@ -1,11 +1,8 @@
 import videojs from 'video.js';
 import {version as VERSION} from '../package.json';
-import { CmcdRequest } from './cmcdKeys/cmcdRequest';
-import { CmcdObject } from './cmcdKeys/cmcdObject';
-import { CmcdSession } from './cmcdKeys/cmcdSession';
-import { CmcdStatus } from './cmcdKeys/cmcdStatus';
+import { appendCmcdHeaders } from '@svta/common-media-library/cmcd/appendCmcdHeaders';
 import { appendCmcdQuery } from '@svta/common-media-library/cmcd/appendCmcdQuery';
-import { keyTypes } from './cmcdKeys/common.js';
+import { uuid } from '@svta/common-media-library/utils/uuid';
 
 const Plugin = videojs.getPlugin('plugin');
 
@@ -41,7 +38,7 @@ class Cmcd extends Plugin {
     const {sid, cid, useHeaders} = options || {};
 
     this.cid = cid;
-    this.sid = sid || generateUuid();
+    this.sid = sid || uuid();
     this.useHeaders = useHeaders;
 
     this.player.ready(() => {
@@ -52,23 +49,11 @@ class Cmcd extends Plugin {
       player.on('xhr-hooks-ready', () => {
 
         const playerXhrRequestHook = (opts) => {
-          const cmcdRequest = new CmcdRequest(this.player);
-          const cmcdObject = new CmcdObject(this.player);
-          const cmcdSession = new CmcdSession(this.player, this.sid, this.cid);
-          const cmcdStatus = new CmcdStatus(this.player);
-
-          const keyObject = cmcdObject.getKeys(opts.uri);
-          const keyRequest = cmcdRequest.getKeys(opts.uri, isWaitingEvent, keyObject.ot);
-          const keySession = cmcdSession.getKeys(this.player.currentSrc());
-          const keyStatus = cmcdStatus.getKeys(isWaitingEvent);
+          const cmcd = new Cmcd(this.player, this.sid, this.cid);
+          const keys = cmcd.getKeys(opts.uri, isWaitingEvent, this.player.currentSrc());
 
           if (this.useHeaders) {
-            const headers = toHeader({
-              'CMCD-Request': keyRequest,
-              'CMCD-Object': keyObject,
-              'CMCD-Session': keySession,
-              'CMCD-Status': keyStatus
-            });
+            const headers = appendCmcdHeaders({}, keys);
 
             opts.beforeSend = (xhr) => {
               for (const h in headers) {
@@ -76,9 +61,7 @@ class Cmcd extends Plugin {
               }
             };
           } else {
-            const cmcdKeysObject = Object.assign({}, keyRequest, keyObject, keySession, keyStatus);
-
-            opts.uri = appendCmcdQuery(opts.uri, cmcdKeysObject);
+            opts.uri = appendCmcdQuery(opts.uri, keys);
           }
 
           return opts;
@@ -100,54 +83,6 @@ class Cmcd extends Plugin {
     }
   }
 
-}
-
-function generateUuid() {
-  const uuid = new Uint8Array(16);
-
-  for (let i = 0; i < uuid.length; i++) {
-    uuid[i] = Math.floor(Math.random() * 256);
-  }
-
-  // Set the version number to 4
-  uuid[6] = (uuid[6] & 0x0f) | 0x40;
-
-  // Set the variant number to 2
-  uuid[8] = (uuid[8] & 0x3f) | 0x80;
-
-  // Convert the UUID to a string in the canonical format
-  const hexDigits = Array.from(uuid, byte => byte.toString(16).padStart(2, '0'));
-  const hexString = hexDigits.join('');
-
-  return `${hexString.slice(0, 8)}-${hexString.slice(8, 12)}-4${hexString.slice(13, 16)}-${String.fromCharCode((uuid[8] & 0x0f) | 0x80)}${hexString.slice(17, 20)}-${hexString.slice(20)}`;
-}
-
-function objToString(obj) {
-  const keys = Object.keys(obj);
-  const keyValPair = keys.map((key) => {
-    if (keyTypes[key] === 'string') {
-      return `${key}="${obj[key]}"`;
-    } else if (keyTypes[key] === 'boolean' && obj[key] === true) {
-      return `${key}`;
-    }
-    return `${key}=${obj[key]}`;
-
-  });
-
-  return keyValPair.filter(s => !s.includes('undefined')).join(',');
-}
-
-function toHeader(cmcdHeaders) {
-  const headers = {};
-
-  for (const header in cmcdHeaders) {
-    const str = objToString(cmcdHeaders[header]);
-
-    if (str.length !== 0) {
-      headers[header] = str;
-    }
-  }
-  return headers;
 }
 
 function handleEvents(player) {
